Use the Timer to trigger updates
I removed the switch statement from the run() method, because the timer
will only call it if the simulation is running. I also removed the
Thread thing from that class, because it is not needed -- a Runnable
object basically IS a thread. Finally, I made the End button call end_sim
(instead of start_sim) and disable the Start button, because the timer
cannot be restarted after it has been terminated.