Add proper line-of-sight checking to Predators
I also created the PredatorTests class to test it, and removed the note
in move_toward about collision detection because that function doesn’t
have access to the Rock list -- it can’t do collision detection.

Also, predator height is currently hard-coded. It matters, because of
the shape of rocks (they are domes, not infinitely tall cylinders), but
I cannot find it in the data file or specification file.